% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{createFeatures}
\alias{createFeatures}
\title{mapedit create features Addin}
\usage{
createFeatures(SF_OBJECT = NULL)
}
\arguments{
\item{SF_OBJECT}{sf Simple feature collection}
}
\value{
sf object and/or saved to file
}
\description{
Create and save spatial objects within the Rstudio IDE. Objects
can then be saved to file types such as \code{.geojson} or \code{.shp}.
Objects are also output to the console and can be assigned to a variable
using \code{.Last.value}. If you wish to pass the output directly to a variable
simply call the addin function, ie. \code{new_sf <- createFeatures()}.

An existing sf \code{data.frame} can also be passed either indirectly by
selecting text in RStudio with the name of the object, or directly by
passing the existing sf object to \code{new_sf <- createFeatures(existing_sf)}.
When passing an existing sf object you can only add and edit additional features,
the existing features cannot be changed.
}
