% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{edit_map}
\alias{edit_map}
\title{Interactively Edit a Map}
\usage{
edit_map(x, ...)
}
\arguments{
\item{x}{map to edit}

\item{...}{other arguments}
}
\description{
Interactively Edit a Map
}
\examples{
library(leaflet)
library(mapedit)
edit_map(leaflet() \%>\% addTiles())

\dontrun{
  # demonstrate Leaflet.Draw on a layer

  # will use the very new randgeo
  # devtools::install_github("ropensci/randgeo")

  library(leaflet)
  library(leaflet.extras)
  library(randgeo)  # see install instructions above

  lf <- leaflet() \%>\%
    addTiles() \%>\%
    addGeoJSON(
      geo_polygon(
        count = 10,
        bbox = c(50,50,60,60),
        max_radial_length = 2
      ),
      layerId = "randgeo"
    ) \%>\%
    addDrawToolbar(
      targetLayer = "randgeo",
      editOptions = editToolbarOptions()
    ) \%>\%
    fitBounds(48,48,62,62)

  library(mapedit)
  edit_map(lf)
}
}
