% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{editModUI}
\alias{editModUI}
\title{Shiny Module UI for Geo Create, Edit, Delete}
\usage{
editModUI(id, ...)
}
\arguments{
\item{id}{\code{character} id for the the Shiny namespace}

\item{...}{other arguments to \code{leafletOutput()}}
}
\value{
ui for Shiny module
}
\description{
Shiny Module UI for Geo Create, Edit, Delete
}
