% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapedit.R
\docType{package}
\name{mapedit-package}
\alias{mapedit}
\alias{mapedit-package}
\title{mapedit: interactive editing and selection for geospatial data}
\description{
mapedit, a RConsortium funded project, provides interactive
tools to incorporate in geospatial workflows that require editing or selection
of spatial data.
}
\section{Edit}{

\itemize{
\item{\code{\link{editMap}}}
\item{\code{\link{editFeatures}}}
\item{Shiny edit module \code{\link{editModUI}}, \code{\link{editMod}}}
}

#' @section Edit:
\itemize{
\item{\code{\link{selectMap}}}
\item{\code{\link{selectFeatures}}}
\item{Shiny edit module \code{\link{selectModUI}}, \code{\link{selectMod}}}
}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-spatial/mapedit}
  \item Report bugs at \url{https://github.com/r-spatial/mapedit/issues}
}

}
\author{
\strong{Maintainer}: Tim Appelhans \email{tim.appelhans@gmail.com}

Authors:
\itemize{
  \item Kenton Russell
  \item Lorenzo Busetto
}

}
