% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin.R
\name{mapeditAddin}
\alias{mapeditAddin}
\title{mapedit Addin}
\usage{
mapeditAddin()
}
\value{
sf object to .GlobalEnv and/or saved to file
}
\description{
Create and save spatial objects within the Rstudio IDE
}
