% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_add}
\alias{merge_add}
\title{Merge 'sf' Adds}
\usage{
merge_add(orig = NULL, drawn = NULL)
}
\arguments{
\item{orig}{\code{sf} with the original or source data to which
adds should apply}

\item{drawn}{\code{sf} with sf data to add to orig}
}
\description{
Internal function used with \code{editFeatures} to apply adds or drawn
  to a \code{sf} object.
}
\keyword{internal}
