% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge_delete}
\alias{merge_delete}
\title{Merge 'sf' Deletes}
\usage{
merge_delete(orig = NULL, deletes = NULL, by = c(id = "layerId"))
}
\arguments{
\item{orig}{\code{sf} with the original or source data to which
deletes should apply}

\item{deletes}{\code{sf} with sf data to delete}

\item{by}{named \code{vector} with the name of the vector representing
the column in orig that we will use to match and the value of
the vector representing the column in deletes that we will
use to match.  The argument is intended to work like
the \code{*join} functions in \code{dplyr}. Note, this function
will only use the first name
and first value of the vector for matching.}
}
\description{
Internal function used with \code{editFeatures} to apply deletes
  to a \code{sf} object.
}
\keyword{internal}
