% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%>\%}
\alias{\%>\%}
\title{Pipe operator}
\usage{
lhs \%>\% rhs
}
\value{
\code{NULL} (this is the magrittr pipe operator)
}
\description{
See \code{magrittr::\link[magrittr:pipe]{\%>\%}} for details.
}
\keyword{internal}
