% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToolbar.R
\name{processOpts}
\alias{processOpts}
\title{Prepare arguments for addPmToolbar}
\usage{
processOpts(fun, args)
}
\arguments{
\item{fun}{Function used by editor package (leafpm) to set defaults}

\item{args}{Either a (possibly nested) list of named options of
the form suitable for passage to \code{fun}.}
}
\value{
An object suitable for passing in as the supplied argument
to \code{leafpm::addPmToolbar}.
}
\description{
Prepare arguments for addPmToolbar
}
