% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{selectMod}
\alias{selectMod}
\title{Shiny Module Server for Geo Selection}
\usage{
selectMod(
  input,
  output,
  session,
  leafmap,
  styleFalse = list(fillOpacity = 0.2, weight = 1, opacity = 0.4),
  styleTrue = list(fillOpacity = 0.7, weight = 3, opacity = 0.7)
)
}
\arguments{
\item{input}{Shiny server function input}

\item{output}{Shiny server function output}

\item{session}{Shiny server function session}

\item{leafmap}{leaflet map to use for Selection}

\item{styleFalse}{named \code{list} of valid \code{CSS} for non-selected features}

\item{styleTrue}{named \code{list} of valid \code{CSS} for selected features}
}
\value{
server function for Shiny module
}
\description{
Shiny Module Server for Geo Selection
}
