% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToolbar.R
\name{addToolbar}
\alias{addToolbar}
\title{Add a (possibly customized) toolbar to a leaflet map}
\usage{
addToolbar(leafmap, editorOptions, editor, targetLayerId)
}
\arguments{
\item{leafmap}{leaflet map to use for Selection}

\item{editorOptions}{A list of options to be passed on to either
\code{leaflet.extras::addDrawToolbar} or
\code{leafpm::addPmToolbar}.}

\item{editor}{Character string giving editor to be used for the
current map. Either \code{"leafpm"} or
\code{"leaflet.extras"}.}

\item{targetLayerId}{\code{string} name of the map layer group to
use with edit}
}
\value{
The leaflet map supplied to \code{leafmap}, now with an
    added toolbar.
}
\description{
Add a (possibly customized) toolbar to a leaflet map
}
