% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edit.R
\name{editMap}
\alias{editMap}
\title{Interactively Edit a Map}
\usage{
editMap(x, ...)
}
\arguments{
\item{x}{map to edit}

\item{...}{other arguments}
}
\description{
Interactively Edit a Map
}
\examples{
library(leaflet)
library(mapedit)
editMap(leaflet() \%>\% addTiles())

\dontrun{
  # demonstrate Leaflet.Draw on a layer
  library(sf)
  library(mapview)
  library(leaflet.extras)
  library(mapedit)

  # ?sf::sf
  pol = st_sfc(
    st_polygon(list(cbind(c(0,3,3,0,0),c(0,0,3,3,0)))),
    crs = 4326
  )
  mapview(pol) \%>\%
    editMap(targetLayerId = "pol")

  mapview(franconia[1:2,]) \%>\%
    editMap(targetLayerId = "franconia[1:2, ]")
}
}
