% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\name{editMod}
\alias{editMod}
\title{Shiny Module Server for Geo Create, Edit, Delete}
\usage{
editMod(input, output, session, leafmap, targetLayerId = NULL, sf = TRUE)
}
\arguments{
\item{input}{Shiny server function input}

\item{output}{Shiny server function output}

\item{session}{Shiny server function session}

\item{leafmap}{leaflet map to use for Selection}

\item{targetLayerId}{\code{character} identifier of layer to edit, delete}

\item{sf}{\code{logical} to return simple features.  \code{sf=FALSE} will return
\code{GeoJSON}.}
}
\value{
server function for Shiny module
}
\description{
Shiny Module Server for Geo Create, Edit, Delete
}
