% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addToolbar.R
\name{processOpts}
\alias{processOpts}
\title{Prepare arguments for addDrawToolbar or addPmToolbar}
\usage{
processOpts(fun, args)
}
\arguments{
\item{fun}{Function used by editor package (leafpm or
leaflet.extras) to set defaults}

\item{args}{Either a (possibly nested) list of named options of
the form suitable for passage to \code{fun} or (if the chosen
editor is \code{"leaflet.extras"}) \code{FALSE}.}
}
\value{
An object suitable for passing in as the supplied argument
    to either \code{leaflet.extras::addDrawToolbar} or
    \code{leafpm::addPmToolbar}.
}
\description{
Prepare arguments for addDrawToolbar or addPmToolbar
}
